Adds season stats module and venue stats module
