+++ test/stat_tracker_team_test.rb
@@ -39,13 +39,13 @@ class StatTrackerTeamTest < Minitest::Test
   def test_it_gets_games_by_team_id
     game = @stat_tracker.games
     expected = [game[0], game[1], game[2], game[3], game[4]]
-    assert_equal expected, @stat_tracker.games_by_team("3")
+    assert_equal expected, @stat_tracker.games_by_team_id("3")
   end
 
   def test_it_gets_games_by_season
     game = @stat_tracker.games
     expected = [game[6], game[7], game[8], game[9]]
-    assert_equal expected, @stat_tracker.games_by_season("20152016", @stat_tracker.games)
+    assert_equal expected, @stat_tracker.games_by_team_by_season("20152016", @stat_tracker.games)
   end
 
   def test_it_gets_games_by_season_for_team_id
